feat: normalized CID (#4)
As discussed, this PR makes the CID the primary key in `content`. It will be normalized as v1 b32. This is prehaps not the most efficient format for strorage/indexing (alternative was `BYTEA`) but is significantly easier to query and display as a `TEXT` field.

It also adds a `cid` column to the `upload` table for recording the CID as the user uploaded it (e.g may be a v0).

fixes https://github.com/web3-storage/web3.storage/issues/438