feat: cargo fdw (#5)
This PR has changes to add `dagcargo` as a **Foreign Data Wrapper**.

Sorry for the noise, I had to separate out the table and type definitions from `CREATE SCHEMA` since it doesn't support `CREATE TYPE` in the `schema_element` list. i.e. there was a syntax error in the current schema that _had_ to be fixed to enable me to run this:

> Currently, only CREATE TABLE, CREATE VIEW, CREATE INDEX, CREATE SEQUENCE, CREATE TRIGGER and GRANT are accepted as clauses within CREATE SCHEMA.
> https://www.postgresql.org/docs/9.3/sql-createschema.html

Anyway, this adds the `deals, aggregates, aggregate_entries` tables to the web3storage schema allowing you to write queries in the web3.storage database like this, that reference tables in the dagcargo database:

```sql
SELECT c.cid, c.dag_size, ae.aggregate_cid, ae.datamodel_selector, d.deal_id, d.status
  FROM web3storage.content AS c
  JOIN web3storage.aggregate_entries AS ae
    ON ae.cid_v1 = c.cid
  JOIN web3storage.aggregates AS a
    ON ae.aggregate_cid = a.aggregate_cid
  JOIN web3storage.deals as d
    ON ae.aggregate_cid = d.aggregate_cid
 WHERE c.cid = 'bafybeidull6drmxinh3fx2viqrgb7rczn7mrqs775xaypx3zqefqpedzkm'
```

We no longer need to store a copy of the dagcargo data in our database, and dagcargo no longer needs to update our database with the information it has.

Aside from semicolons and schema prefixes the real change to `schema.sql` is the deletion of the `aggregate`, `aggregate_entry` and `deal` tables.

More info on fdw: https://www.postgresql.org/docs/10/postgres-fdw.html

* [x] depends on https://github.com/web3-storage/web3-schema/pull/4